requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.12.0
gunicorn==21.2.0

